+++ ugbase/lib_algebra/martin_algebra/SparseMatrix.h
@@ -55,7 +55,7 @@ public:
 
 	struct connection
 	{
-		int iIndex;		// index to
+		size_t iIndex;		// index to
 		entry_type dValue; // smallmatrix value;
 		
 		void print(){cout << *this;}
@@ -87,7 +87,7 @@ public: // construction etc
 	~SparseMatrix ();	
 	
 
-	bool create(int _rows, int _cols);
+	bool create(size_t _rows, size_t _cols);
 	bool destroy();
 	
 	//! create this as a transpose of SparseMatrix B
@@ -102,8 +102,8 @@ public:	// general functions
 	template<typename Vector_type>
 	void eliminateDirichletValues(Vector_type &b);
 	
-	void setDirichletRow(int row);
-	void setDirichletRows(int *pRows, int nrows);
+	void setDirichletRow(size_t row);
+	void setDirichletRows(size_t *pRows, size_t nrows);
 	bool set_dirichlet_rows(const local_index_type &ind);
 	
 	//! calculate res = A x
@@ -119,41 +119,41 @@ public:	// general functions
 	bool matmul_minus(Vector_type &res, const Vector_type &x) const;
 
 	//! accessor functions for artificial matrixrow-object (= just wrapper with A and row)	
-	inline const matrixrow_type getrow(int i) const;
-	const matrixrow_type operator [] (int i) const;
+	inline const matrixrow_type getrow(size_t i) const;
+	const matrixrow_type operator [] (size_t i) const;
 	
 	
 	//! get Diagonal A_[i,i] of matrix
-	inline const entry_type &getDiag(int i) const;
-	inline entry_type &getDiag(int i);
+	inline const entry_type &getDiag(size_t i) const;
+	inline entry_type &getDiag(size_t i);
 	
 	//! isUnconnected: true if only A[i,i] != 0.0.
-	inline bool isUnconnected(int i) const;
+	inline bool isUnconnected(size_t i) const;
 
 	//! adds the submatrix mat to A. 	
 	
 	template<typename M>
-	void add(const M &mat, int *rows, int *cols);
+	void add(const M &mat, size_t *rows, size_t *cols);
 	template<typename M>
-	void set(const M &mat, int *rows, int *cols);
+	void set(const M &mat, size_t *rows, size_t *cols);
 	template<typename M>
-	void get(M &mat, int *rows, int *cols) const;
+	void get(M &mat, size_t *rows, size_t *cols) const;
 	
 	template<typename M>
-	void add(const M &mat, vector<int> &rows, vector<int> &cols);
+	void add(const M &mat, vector<size_t> &rows, vector<size_t> &cols);
 	template<typename M>
-	void set(const M &mat, vector<int> &rows, vector<int> &cols);
+	void set(const M &mat, vector<size_t> &rows, vector<size_t> &cols);
 	template<typename M>
-	void get(M &mat, vector<int> &rows, vector<int> &cols) const;
+	void get(M &mat, vector<size_t> &rows, vector<size_t> &cols) const;
 	
 	bool add(const local_matrix_type &mat, const local_index_type &I, const local_index_type &J);
 	bool set(const local_matrix_type &mat, const local_index_type &I, const local_index_type &J);
 	bool get(local_matrix_type &mat, const local_index_type &I, const local_index_type &J) const;
 	
 	
-	void add(const entry_type &d, int row, int col);
-	void set(const entry_type &d, int row, int col);
-	void get(entry_type &d, int row, int col) const;
+	void add(const entry_type &d, size_t row, size_t col);
+	void set(const entry_type &d, size_t row, size_t col);
+	void get(entry_type &d, size_t row, size_t col) const;
 	
 	bool set(double a);
 
@@ -163,31 +163,31 @@ public:	// general functions
 		
 public: // accessor functions
 	
-	int getLength() const { return rows; }
-	int getRows() const { return rows; }
-	int getCols() const { return cols; }
-	int row_size() const { return rows; }
-	int col_size() const { return rows; }
+	size_t getLength() const { return rows; }
+	size_t getRows() const { return rows; }
+	size_t getCols() const { return cols; }
+	size_t row_size() const { return rows; }
+	size_t col_size() const { return rows; }
 	
-	int getTotalNrOfConnections() const { return iTotalNrOfConnections; }
+	size_t getTotalNrOfConnections() const { return iTotalNrOfConnections; }
 	
 	
 	
 public:	// row functions
 	//! remove zero entries of SparseMatrix (experimental)
-	void removezeros(int row);
+	void removezeros(size_t row);
 	
-	void setMatrixRow(int row, connection *c, int nr);
-	void addMatrixRow(int row, connection *c, int nr);
-	int getNrOfConnections(int row) const { return iNrOfConnections[row]; }
+	void setMatrixRow(size_t row, connection *c, size_t nr);
+	void addMatrixRow(size_t row, connection *c, size_t nr);
+	size_t getNrOfConnections(size_t row) const { return iNrOfConnections[row]; }
 
 public: // output functions
 
 	void print(const char * const name = NULL) const;
 	void printToFile(const char *filename) const;
-	void printrow(int row) const;
+	void printrow(size_t row) const;
 	void p() const; // for use in gdb
-	void pr(int i) const; // for use in gdb
+	void pr(size_t i) const; // for use in gdb
 
 	friend ostream &operator<<(ostream &output, const SparseMatrix &m)
 	{
@@ -217,8 +217,8 @@ public:
 	{	
 	public:
 		diagcomponent(const SparseMatrix &A_) : A(A_) {}
-		entry_type operator [] (int i) const { return A.getDiag(i); }
-		int getLength() const { return A.getLength(); }
+		entry_type operator [] (size_t i) const { return A.getDiag(i); }
+		size_t getLength() const { return A.getLength(); }
 		friend ostream &operator<<(ostream &output, const diagcomponent &m)
 		{ output << "diagonal of " << m.A; return output; }
 		const SparseMatrix &A;
@@ -230,7 +230,7 @@ public:
 	}
 	
 	//! prefetch (EXPERIMENTAL)
-	inline void prefetch(int i) const
+	inline void prefetch(size_t i) const
 	{
 		prefetchRead(cons[i]);
 	}
@@ -247,10 +247,10 @@ public:
 	public:
 		//const SparseMatrix<entry_type> &A;
 		const SparseMatrix<entry_type> *A;
-		int m_position;
-		int row;
+		size_t m_position;
+		size_t row;
 	public:
-		inline cRowIterator(const SparseMatrix<entry_type> &A_, int row_)
+		inline cRowIterator(const SparseMatrix<entry_type> &A_, size_t row_)
 		{
 			A = &A_;
 			row = row_; 
@@ -264,7 +264,7 @@ public:
 		inline void operator ++() {	m_position++; }
 		
 		inline void rewind() { m_position = 0;}
-		inline int getPos() const{	return m_position;}
+		inline size_t getPos() const{	return m_position;}
 		
 		inline bool isEnd() const { return m_position >= A->getNrOfConnections(row); }
 	};
@@ -277,9 +277,9 @@ public:
 		const connection * const pEnd;
 		const connection * const pStart;
 		const connection *p;
-		const int row;
+		const size_t row;
 	public:
-		inline cRowIterator(const SparseMatrix<entry_type> &A, int row_) :
+		inline cRowIterator(const SparseMatrix<entry_type> &A, size_t row_) :
 			pStart(A.cons[row_]), 
 			pEnd( A.cons[row_]+A.getNrOfConnections(row_) ),
 			row(row_)
@@ -296,7 +296,7 @@ public:
 		inline void operator ++() {	p++; }
 		
 		inline void rewind() { p = pStart;}
-		inline int getPos() const{	return p-pStart;}
+		inline size_t getPos() const{	return p-pStart;}
 		
 		inline bool isEnd() const { return p >= pEnd; }
 	};*/
@@ -306,11 +306,11 @@ class cRowIterator
 public:
 	//const SparseMatrix<entry_type> &A;
 	const connection * const pStart;
-	const int nrOfConnections;
-	const int row;
-	int pos;
+	const size_t nrOfConnections;
+	const size_t row;
+	size_t pos;
 public:
-	inline cRowIterator(const SparseMatrix<entry_type> &A, int row_) :	pStart(A.cons[row_]), nrOfConnections( A.getNrOfConnections(row_) ),
+	inline cRowIterator(const SparseMatrix<entry_type> &A, size_t row_) :	pStart(A.cons[row_]), nrOfConnections( A.getNrOfConnections(row_) ),
 		row(row_)
 	{
 		rewind();
@@ -324,13 +324,13 @@ public:
 	inline void operator ++() {	pos++; }
 
 	inline void rewind() { pos = 0;}
-	inline int getPos() const{	return pos;}
+	inline size_t getPos() const{	return pos;}
 	
 	inline bool isEnd() const { return pos >= nrOfConnections; }
 };
 #endif
 	
-	cRowIterator beginRow(int row) const
+	cRowIterator beginRow(size_t row) const
 	{
 		return cRowIterator(*this, row);
 	}
@@ -341,10 +341,10 @@ public:
 	{
 	public:
 		SparseMatrix<entry_type> &A;
-		int m_position;
-		int row;
+		size_t m_position;
+		size_t row;
 	public:
-		inline rowIterator(SparseMatrix<entry_type> &A_, int row_) : A(A_) { row = row_; rewind(); }
+		inline rowIterator(SparseMatrix<entry_type> &A_, size_t row_) : A(A_) { row = row_; rewind(); }
 		inline rowIterator(const cRowIterator &other) : A(other.A) { row = other.row; m_position = other.m_position; }
 		
 		inline connection &operator *() const {return A.cons[row][m_position];}
@@ -352,12 +352,12 @@ public:
 		inline void operator ++() {	m_position++; }
 		
 		inline void rewind() { m_position = 0;}
-		inline int getPos() const {	return m_position;}
+		inline size_t getPos() const {	return m_position;}
 		
 		inline bool isEnd() const { return m_position >= A.getNrOfConnections(row); }
 	};
 	
-	rowIterator beginRow(int row)
+	rowIterator beginRow(size_t row)
 	{
 		return rowIterator(*this, row);
 	}
@@ -368,13 +368,13 @@ public:
 	{
 	protected:
 		matrixrow<entry_type> &row;
-		int m_position;
+		size_t m_position;
 	public:
 		inline iterator(matrixrow<entry_type> &r) : row(r) { m_position = 0;}
 		inline connection &operator *(){	return row[m_position];	}
 		inline void operator ++(){		m_position++;			}
 		inline void rewind(){m_position = 0;}
-		inline int getPos() const{	return m_position;	}
+		inline size_t getPos() const{	return m_position;	}
 		inline const matrixrow &getMatrixRow() const{return row;}
 		inline bool isEnd() const { return m_position >= row.getNrOfConnections(); }
 	}
@@ -384,7 +384,7 @@ public:
 	class cLowerLeftIterator : public cRowIterator
 	{
 	public:
-		cLowerLeftIterator(const SparseMatrix<entry_type> &A_, int row_) : cRowIterator(A_, row_) { 	rewind();	}
+		cLowerLeftIterator(const SparseMatrix<entry_type> &A_, size_t row_) : cRowIterator(A_, row_) { 	rewind();	}
 		inline void operator ++()
 		{
 			cRowIterator::m_position ++;
@@ -399,7 +399,7 @@ public:
 		}
 	};
 				 
-	cLowerLeftIterator beginLowerLeftRow(int row)  const
+	cLowerLeftIterator beginLowerLeftRow(size_t row)  const
 	{
 		return cLowerLeftIterator(*this, row);
 	}
@@ -407,7 +407,7 @@ public:
 	class cUpperRightIterator : public cRowIterator
 	{
 	public:
-		cUpperRightIterator(const SparseMatrix<entry_type> &A_, int row_) : cRowIterator(A_, row_) { 	rewind();	}
+		cUpperRightIterator(const SparseMatrix<entry_type> &A_, size_t row_) : cRowIterator(A_, row_) { 	rewind();	}
 		inline void operator ++()
 		{
 			cRowIterator::m_position ++;
@@ -423,7 +423,7 @@ public:
 	};
 				
 				
-	cUpperRightIterator beginUpperRightRow(int row)  const
+	cUpperRightIterator beginUpperRightRow(size_t row)  const
 	{
 		return cUpperRightIterator(*this, row);
 	}
@@ -431,7 +431,7 @@ public:
 	class cLowerLeftIterator : public cRowIterator
 	{
 	public:
-		cLowerLeftIterator(const SparseMatrix<entry_type> &A, int row_) : cRowIterator(A, row_) { 	rewind();	}
+		cLowerLeftIterator(const SparseMatrix<entry_type> &A, size_t row_) : cRowIterator(A, row_) { 	rewind();	}
 		inline void operator ++()
 		{
 			cRowIterator::operator ++();
@@ -448,7 +448,7 @@ public:
 		}
 	};
 	
-	cLowerLeftIterator beginLowerLeftRow(int row)  const
+	cLowerLeftIterator beginLowerLeftRow(size_t row)  const
 	{
 		return cLowerLeftIterator(*this, row);
 	}
@@ -456,7 +456,7 @@ public:
 	class cUpperRightIterator : public cRowIterator
 		{
 		public:
-			cUpperRightIterator(const SparseMatrix<entry_type> &A, int row_) : cRowIterator(A, row_) { 	rewind();	}
+			cUpperRightIterator(const SparseMatrix<entry_type> &A, size_t row_) : cRowIterator(A, row_) { 	rewind();	}
 			inline void operator ++()
 			{
 				cRowIterator::operator ++();
@@ -473,7 +473,7 @@ public:
 		};
 	
 	
-	cUpperRightIterator beginUpperRightRow(int row)  const
+	cUpperRightIterator beginUpperRightRow(size_t row)  const
 	{
 		return cUpperRightIterator(*this, row);
 	}
@@ -482,11 +482,11 @@ public:
 	
 private:	
 	//! "safe" way to set a connection, since when cons[row] is in the big consecutive consmem-array,
-	void safeSetConnections(int row, connection *mem) const;
-	void recreateWithMaxNrOfConnections(int newMax) const;
+	void safeSetConnections(size_t row, connection *mem) const;
+	void recreateWithMaxNrOfConnections(size_t newMax) const;
 	
 public:
-	void setEstimatedRowSize(int estimatedRowSize_)
+	void setEstimatedRowSize(size_t estimatedRowSize_)
 	{
 		estimatedRowSize = estimatedRowSize_;
 	}
@@ -494,27 +494,27 @@ public:
 //----------------
 	
 public:
-	int fromlevel;						//!< fromlevel
-	int tolevel;						//!< tolevel
+	size_t fromlevel;						//!< fromlevel
+	size_t tolevel;						//!< tolevel
 	const char *name;					//!< name of the SparseMatrix for debuging / printing.
 	
 private:	
 
-	int rows;							//!< nr of rows
-	int cols;							//!< nr of cols
-	int *iNrOfConnections;				//< array nr of Connections for each row
+	size_t rows;							//!< nr of rows
+	size_t cols;							//!< nr of cols
+	size_t *iNrOfConnections;				//< array nr of Connections for each row
 	connection **cons;			//< pointers to array of connections of each row
 	
-	int iTotalNrOfConnections;	//!< number of non-zeros
-	int bandwidth;				//!< bandwidth (experimental)
+	size_t iTotalNrOfConnections;	//!< number of non-zeros
+	size_t bandwidth;				//!< bandwidth (experimental)
 	
-	int estimatedRowSize;				//!< estimated length of each row
-	int *iMaxNrOfConnections;			//!< max nr of connections for row [i]. TODO.
+	size_t estimatedRowSize;				//!< estimated length of each row
+	size_t *iMaxNrOfConnections;			//!< max nr of connections for row [i]. TODO.
 
 	
 	connection *consmem;		//!< consecutive memory for the connections
-	int consmemsize;					//!< size of the consecutive memory for connections
-	int iFragmentedMem;			//!< size of connections memory not in consmem
+	size_t consmemsize;					//!< size of the consecutive memory for connections
+	size_t iFragmentedMem;			//!< size of connections memory not in consmem
 	
 	friend class matrixrow<entry_type>;
